requirements.txt
gunicorn
Flask
Flask-Cors
Flask-Script
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Bcrypt
Flask-Caching
Flask-Limiter
Flask-WTF
Flask-Marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
redis
gunicorn
